number-theory

+++ number-theory/mod/mod-pow.cpp
@@ -1,5 +1,5 @@
-int mod_pow(int128 a, int e, const int mod) {
-    int128 r = 1;
+int mod_pow(int a, int e, const int mod) {
+    int r = 1;
     while (e != 0) {
         if (e & 1) r = (r * a) % mod;
         a = (a * a) % mod;
